+++ api-client/src/index.ts
@@ -6,6 +6,7 @@ export * from './deck_configuration'
 export * from './errorRecovery'
 export * from './health'
 export * from './instruments'
+export * from './labware_offsets'
 export * from './maintenance_runs'
 export * from './modules'
 export * from './networking'
