+++ app/src/organisms/LabwarePositionCheck/hooks/useLPCCommands/useApplyLPCOffsets.ts
@@ -1,6 +1,6 @@
 import { useState } from 'react'
 
-import { useCreateLabwareOffsetMutation } from '@opentrons/react-api-client'
+import { useCreateLegacyLabwareOffsetMutation } from '@opentrons/react-api-client'
 
 import type { LegacyLabwareOffsetCreateData } from '@opentrons/api-client'
 import type { UseLPCCommandChildProps } from './types'
@@ -23,7 +23,7 @@ export function useApplyLPCOffsets({
 }: UseApplyLPCOffsetsProps): UseApplyLPCOffsetsResult {
   const [isApplyingOffsets, setIsApplyingOffsets] = useState<boolean>(false)
 
-  const { createLabwareOffset } = useCreateLabwareOffsetMutation()
+  const { createLabwareOffset } = useCreateLegacyLabwareOffsetMutation()
 
   const handleApplyOffsetsAndClose = (
     offsets: LegacyLabwareOffsetCreateData[]
