+++ react-api-client/src/runs/__tests__/useCreateLegacyLabwareOffsetsMutation.test.tsx
@@ -2,10 +2,10 @@ import type * as React from 'react'
 import { describe, it, expect, beforeEach, vi } from 'vitest'
 import { QueryClient, QueryClientProvider } from 'react-query'
 import { act, renderHook, waitFor } from '@testing-library/react'
-import { createLabwareOffset } from '@opentrons/api-client'
+import { createLegacyLabwareOffset } from '@opentrons/api-client'
 import { useHost } from '../../api'
 
-import { useCreateLabwareOffsetMutation } from '../useCreateLabwareOffsetMutation'
+import { useCreateLegacyLabwareOffsetMutation } from '../useCreateLegacyLabwareOffsetMutation'
 import type {
   HostConfig,
   LegacyLabwareOffsetCreateData,
@@ -20,7 +20,7 @@ const DEFINITION_URI = 'definition_uri'
 const LABWARE_LOCATION = { slotName: '4' }
 const OFFSET = { x: 1, y: 2, z: 3 }
 
-describe('useCreateLabwareOffsetMutation hook', () => {
+describe('useCreateLegacyLabwareOffsetMutation hook', () => {
   let wrapper: React.FunctionComponent<{ children: React.ReactNode }>
   let labwareOffset: LegacyLabwareOffsetCreateData
 
@@ -41,11 +41,11 @@ describe('useCreateLabwareOffsetMutation hook', () => {
 
   it('should create labware offsets when callback is called', async () => {
     vi.mocked(useHost).mockReturnValue(HOST_CONFIG)
-    vi.mocked(createLabwareOffset).mockResolvedValue({
+    vi.mocked(createLegacyLabwareOffset).mockResolvedValue({
       data: 'created offsets!',
     } as any)
 
-    const { result } = renderHook(useCreateLabwareOffsetMutation, {
+    const { result } = renderHook(useCreateLegacyLabwareOffsetMutation, {
       wrapper,
     })
