+++ react-api-client/src/runs/index.ts
@@ -16,7 +16,7 @@ export { useAllCommandsQuery } from './useAllCommandsQuery'
 export { useAllCommandsAsPreSerializedList } from './useAllCommandsAsPreSerializedList'
 export { useCommandQuery } from './useCommandQuery'
 export { useRunCommandErrors } from './useRunCommandErrors'
-export * from './useCreateLabwareOffsetMutation'
+export * from './useCreateLegacyLabwareOffsetMutation'
 export * from './useCreateLabwareDefinitionMutation'
 export * from './useUpdateErrorRecoveryPolicy'
 export * from './useErrorRecoveryPolicy'
